Reynold's comments
